Move stdin processing to yield()
Before this patch, stdin is only read between calls to loop(). This
causes problems when a program waits for serial input while calling
yield() in a loop. Moving the stdin handling code to yield() fixes this
problem and more closely matches real hardware behavior.

Also, this patch checks the return value from read(), which fixes a
compiler warning and allows null bytes to be read correctly.